Improved project strcuture + DI with Koin
