+++ src/main/kotlin/Constants.kt
@@ -1,0 +1,3 @@
+import discord4j.common.util.Snowflake
+
+val guildId: Snowflake = Snowflake.of(System.getenv("GUILD_ID"))
