+++ src/main/kotlin/utils/utils.kt
@@ -2,4 +2,7 @@ package utils
 
 import java.util.*
 
+/**
+ * Convert Java Optional to Kotlin nullable
+ */
 fun <T> Optional<T>.getOrNull(): T? = orElse(null)
