+++ src/main/resources/application.conf
@@ -1,0 +1,9 @@
+ktor {
+    deployment {
+        port = 8080
+        port = ${?PORT}
+    }
+    application {
+        modules = [ ApplicationKt.module ]
+    }
+}
